Add internal messaging, DLP deadline tracking, and DSR export flows; expand calendar, notifications, and award UI
- New internal messages system (conversations, directory, unread tracking)
- DLP deadline admin page and data-subject-request send/summary endpoints
- Tender acknowledgement endpoint and agreement acknowledgement modal
- Calendar upcoming view and event detail modal
- Dashboard/admin awards and notifications UI overhaul
- BQ compare/my pages, tender pages, and shared DatePicker/DropdownActions updates

+++ src/app/api/admin/tenders/[id]/bqs/route.ts
@@ -77,7 +77,7 @@ export async function GET(
         ts.contractor_id,
         u.username AS contractor_username,
         COALESCE((
-          SELECT COUNT(*) FROM bq_line_item WHERE submission_id = ts.submission_id
+          SELECT COUNT(*)::int FROM bq_line_item WHERE submission_id = ts.submission_id
         ), 0) AS line_item_count,
         COALESCE((
           SELECT SUM(total_price) FROM bq_line_item WHERE submission_id = ts.submission_id
